feat(Data):  introduce DataStore dependency and implement GameDataStore interface
This commit adds the DataStore dependency to the project and implements the GameDataStore interface. The GameDataStore interface provides methods for fetching and saving all game statistics, as well as saving game data. This addition enables efficient and persistent storage of game-related data, enhancing the overall data management capabilities of the application.